Fix connecting to firestore and drawing data
- added firebase emulators
- added entrypoint & launch configurations for Remote Dev Tools
- fixed missing code that plumbs the database data stream
- added classBoxes to AppState, with corresponding reducer
- fixed StoreClassBoxesMiddleware and DatabaseService to connect
  data stream when action is dispatched
- changed DrawingPage to a StoreConnector above a DrawingCanvas
(with everything that was in DrawingPage)
