+++ .vscode/project.code-snippets
@@ -74,7 +74,7 @@
   "reducer class": {
       "prefix": "reducer",
       "body": [
-          "import 'package:${WORKSPACE_NAME}/models/app_state/app_state.dart';",
+          "import 'package:${WORKSPACE_NAME}/models/app-state/app_state.dart';",
           "import 'package:redux/redux.dart';",
           "",
           "class ${TM_FILENAME_BASE/(.*)/${1:/pascalcase}/g}Reducer extends TypedReducer<AppState, ${TM_FILENAME_BASE/(.*)/${1:/pascalcase}/g}Action> {",
