+++ .vscode/settings.json
@@ -1,0 +1,6 @@
+{
+  "files.exclude": {
+    "**/*.freezed.dart": true,
+    "**/*.g.dart": true
+  }
+}
