+++ firestore.rules
@@ -1,0 +1,12 @@
+rules_version = '2';
+service cloud.firestore {
+  match /databases/{database}/documents {
+    match /{document=**} {
+      allow read, write: if false;
+    }
+    
+    match /domain-objects/{objectId} {
+      allow read, write: if request.auth.uid != null;
+    }
+  }
+}
