+++ lib/actions/shared/connect_data_stream_action.dart
@@ -7,7 +7,7 @@ part 'connect_data_stream_action.g.dart';
 
 @freezed
 class ConnectDataStreamAction with _$ConnectDataStreamAction, ReduxAction {
-  factory ConnectDataStreamAction({required DatabaseSectionEnum section}) =
+  factory ConnectDataStreamAction(DatabaseSectionEnum section) =
       _ConnectDatabaseAction;
 
   factory ConnectDataStreamAction.fromJson(Map<String, dynamic> json) =>
