+++ lib/main_rdt.dart
@@ -1,0 +1,34 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:domain_visualiser/models/app-state/app_state.dart';
+import 'package:domain_visualiser/utils/redux_bundle.dart';
+import 'package:domain_visualiser/utils/store_operation.dart';
+import 'package:domain_visualiser/widgets/app-init/app_widget.dart';
+import 'package:flutter/material.dart';
+import 'package:redux_remote_devtools/redux_remote_devtools.dart';
+
+void main() {
+  // Create the rdt middleware that connects to the rdt server.
+  final _rdtMiddleware = RemoteDevToolsMiddleware<AppState>('localhost:8000');
+
+  // Create an operation for the redux bundle to run on the store.
+  // In this case, we give RemoteDevToolsMiddleware access to the store then
+  // tell it to connect to the server.
+  final _rdtOperation = StoreOperation((store) async {
+    _rdtMiddleware.store = store;
+    await _rdtMiddleware.connect();
+  });
+
+  // Settings to make the firestore package use the local emulator
+  final _firestoreSettings = Settings(
+      host: 'localhost:8080', sslEnabled: false, persistenceEnabled: false);
+
+  // Setup the redux bundle to use a different bucket and with an extra
+  // middleware that sends each action and state to the rdt server for display.
+  ReduxBundle.setup(
+      bucketName: 'gs://profile-pics-prototyping',
+      extraMiddlewares: [_rdtMiddleware],
+      storeOperations: [_rdtOperation],
+      firestoreSettings: _firestoreSettings);
+
+  runApp(AppWidget());
+}
