+++ lib/models/app-state/app_state.dart
@@ -1,5 +1,6 @@
 import 'package:domain_visualiser/enums/auth/auth_step_enum.dart';
 import 'package:domain_visualiser/models/auth/auth_user_data.dart';
+import 'package:domain_visualiser/models/domain-objects/class_box.dart';
 import 'package:domain_visualiser/models/navigation/page_data/page_data.dart';
 import 'package:domain_visualiser/models/problems/problem.dart';
 import 'package:domain_visualiser/models/profile/profile_data.dart';
@@ -18,6 +19,9 @@ class AppState with _$AppState {
     required AuthStepEnum authStep,
     AuthUserData? authUserData,
 
+    /// Domain Objects
+    required IList<ClassBox> classBoxes,
+
     /// Navigation
     required IList<PageData> pagesData,
 
@@ -35,8 +39,9 @@ class AppState with _$AppState {
       _$AppStateFromJson(json);
 
   factory AppState.init() => AppState(
+      authStep: AuthStepEnum.checking,
+      classBoxes: <ClassBox>[].lock,
       pagesData: <PageData>[InitialPageData()].lock,
       problems: IList(),
-      authStep: AuthStepEnum.checking,
       settings: Settings.init());
 }
