+++ test/data/emulators/firebase-export-metadata.json
@@ -1,0 +1,12 @@
+{
+  "version": "9.10.0",
+  "firestore": {
+    "version": "1.11.14",
+    "path": "firestore_export",
+    "metadata_file": "firestore_export/firestore_export.overall_export_metadata"
+  },
+  "auth": {
+    "version": "9.10.0",
+    "path": "auth_export"
+  }
+}
